requirements.txt
Pillow
pyscreenshot
pytesseract
pywin32
pynput
opencv-python
pyautogui
